backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
pdfminer.six>=20221105
pdfplumber>=0.10.0
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8
pandas>=2.0.0